v0.3.0
- removed acf pro repeater field from "about us" section
- minor improvements from zhp.pl version
