+++ .github/workflows/build.yml
@@ -21,6 +21,8 @@ jobs:
       - name: Set current version
         run: echo "VERSION=$(date '+%Y.%m.%d.%H.%M.%S')" >> $GITHUB_ENV
       - uses: actions/checkout@v3
+        with:
+          lfs: true
       - name: Build the Docker image
         run: |
           docker build \
