+++ .github/workflows/zhp-pl.yml
@@ -11,17 +11,12 @@ jobs:
     name: Build
     runs-on: ubuntu-latest
     steps:
-    - name: Set up Docker Buildx
-      uses: docker/setup-buildx-action@v1
-    
     - name: Login to DockerHub
       uses: docker/login-action@v1 
       with:
         username: zhpadmin
         password: ${{ secrets.DOCKERHUB_TOKEN }}
 
-# TODO improve caching in Dockerfile and use github cache to speed up build: https://github.com/docker/build-push-action/blob/master/docs/advanced/cache.md#github-cache
-# possibly, we could wait for this to be implemented: https://github.com/docker/buildx/pull/535
     - name: Build docker
       uses: docker/build-push-action@v2
       with:
@@ -29,17 +24,24 @@ jobs:
         tags: |
           gkzhp/zhp-pl:latest
           gkzhp/zhp-pl:${{ github.sha	}}
+# based on https://blog.benoitblanchon.fr/github-action-run-ssh-commands/
+  deploy-dev:
+    name: Deploy DEV
+    runs-on: ubuntu-latest
+    if: success() && github.ref == 'refs/heads/master'
+    needs: build
+    env:
+      IMAGE_TAG: gkzhp/zhp-pl:${{ github.sha }}
+      CONTAINER_NAME: zhp-pl-dev
+    steps:
+      - name: Configure SSH
+        run: ${{ secrets.ZHP_PL_SSH_CONFIG }}
 
-# TODO prepare docker-compose with the specified tag (git SHA) and env name. Choose proper env file in docker
+      - name: Pull the image
+        run: ssh zhp-pl "docker pull $IMAGE_TAG"
 
-# TODO deploy the docker image,
-  # deploy:
-  #   name: Deploy
-  #   runs-on: ubuntu-latest
-  #   if: success() && github.ref == 'refs/heads/master'
-  #   needs: build
-  #   steps:
-  #   - name: Download package
-  #     uses: actions/download-artifact@v2
-  #     with:
-  #       name: App
+      - name: Remove existing container
+        run: ssh zhp-pl "docker container rm -f $CONTAINER_NAME"
+
+      - name: Start the server
+        run: ssh zhp-pl "docker container run -p 4000:3000 -d --name $CONTAINER_NAME $IMAGE_TAG"
