+++ packages/app/.env
@@ -1,1 +1,2 @@
 REST_API=/wp-json/wp/v2
+SSR=false
