+++ packages/app/package.json
@@ -2,7 +2,7 @@
   "name": "@zhp-pl/app",
   "private": true,
   "license": "MIT",
-  "version": "0.2.3",
+  "version": "0.2.5",
   "scripts": {
     "dev": "nuxt --hostname 0.0.0.0 --post 3000",
     "build": "nuxt build",
@@ -18,7 +18,6 @@
     "@nuxtjs/google-analytics": "^2.4.0",
     "@nuxtjs/pwa": "^3.3.5",
     "@nuxtjs/sentry": "^5.0.3",
-    "@zhp-pl/ui": "^0.2.1",
     "body-scroll-lock": "^3.1.5",
     "lodash.debounce": "^4.0.8",
     "nuxt": "^2.15.6",
