+++ packages/app/router.scrollBehavior.js
@@ -1,3 +1,4 @@
 export default function (to, from, savedPosition) {
-  return { x: 0, y: 0 }
-}
+    return { x: 0, y: 0 }
+  }
+
