+++ packages/ui/src/components/organisms/ZFacebook/ZFacebook.vue
@@ -1,15 +1,16 @@
 <template>
-  <div class="z-facebook">
+  <div class="zb-page z-facebook">
     <div
       class="fb-page"
-      data-href="https://www.facebook.com/ZHPpl"
+      :data-href="fbPageUrl"
       data-tabs="timeline"
       data-width="260"
-      data-height="446"
+      :data-height="height"
       data-small-header="false"
       data-adapt-container-width="true"
       data-hide-cover="false"
       data-show-facepile="true"
+      data-lazy="true"
     />
   </div>
 </template>
@@ -22,6 +23,18 @@ export default {
       type: String,
       default: 'div',
     },
+    appId: {
+      type: String,
+      default: ''
+    },
+    fbPageUrl: {
+      type: String,
+      default: ''
+    },
+    height: {
+      type: Number,
+      default: ''
+    }
   },
   mounted() {
     const installFacebookSdkScript = (d, s, id) => {
@@ -38,7 +51,7 @@ export default {
     installFacebookSdkScript(document, 'script', 'facebook-jssdk');
     window.fbAsyncInit = () => {
       FB.init({
-        appId: '2797653613806518',
+        appId: this.appId,
         cookie: true,
         xfbml: true,
         version: 'v3.2',
@@ -47,8 +60,3 @@ export default {
   },
 };
 </script>
-
-<style lang="scss">
-.z-facebook {
-}
-</style>
